Move zone specific code to views_zones.py.
